+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (anti_build module) fix: the `place_block` anti-type should not throw `NPE` if a non-player (e.g. a dispenser) places a block. (Thanks to @phaldan)
-- i18n: new translation for `zh_tw` and `id_id`. (Thanks to @dirtTW)
+- fix: (chat.display module) The sender player is wrong when sharing `inv` or `ender`.